Added SSLConfigurator, HttpClientConfig

+++ core/src/main/kotlin/io/slychat/messenger/core/http/HttpClientConfig.kt
@@ -1,6 +1,6 @@
 package io.slychat.messenger.core.http
 
 data class HttpClientConfig(
-    val connectTimeoutMs: Long,
-    val readTimeoutMs: Long
+    val connectTimeoutMs: Int,
+    val readTimeoutMs: Int
 )
